+++ src/App.tsx
@@ -4,7 +4,7 @@ import { Helmet, HelmetProvider } from 'react-helmet-async'
 import { RouterProvider } from 'react-router-dom'
 import { Toaster } from 'sonner'
 
-import { ThemeProvider } from './components/theme/theme-provider'
+import { ThemeProvider } from './components/theme/themeProvider'
 import { router } from './routes'
 
 export function App() {
