+++ src/components/accountMenu.tsx
@@ -8,7 +8,7 @@ import {
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from './ui/dropdown-menu'
+} from './ui/dropdownMenu'
 
 export function AccountMenu() {
   return (
